运行路径初始

+++ examples/grpc/app/repository/entity/Entity.go
@@ -3,8 +3,8 @@ package entity
 type entity struct {
 	CreatedAt  string `json:"created_at"`  // 创建时间
 	UpdatedAt  string `json:"updated_at"`  // 更新时间
-	Sort       int    `json:"sort"`        // 排序
-	DataStatus int    `json:"data_status"` // 数据状态
+	Sort       uint32 `json:"sort"`        // 排序
+	DataStatus uint8  `json:"data_status"` // 数据状态
 }
 
 func (e entity) GetCreatedAt() string {
@@ -15,10 +15,10 @@ func (e entity) GetUpdatedAt() string {
 	return e.UpdatedAt
 }
 
-func (e entity) GetSort() int {
+func (e entity) GetSort() uint32 {
 	return e.Sort
 }
 
-func (e entity) GetDataStatus() int {
+func (e entity) GetDataStatus() uint8 {
 	return e.DataStatus
 }
